backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
openai==1.54.0
httpx==0.27.2
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
